Cast size_t in uint in order to avoid a warning.
git-svn-id: file:///Users/matti/Desktop/tmp/procsi/svn/trunk@19 1e18e3dc-80d6-465e-b683-59faa57c043a

+++ src/util.c
@@ -14,7 +14,7 @@ bool readLine(char *str, size_t length)
     // construit la chaîne de formatage de l'entrée
     size_t t_format = 81;
     char *format = calloc(t_format, sizeof(*format));
-    snprintf(format, t_format, "%%%d[^\n]", length);
+    snprintf(format, t_format, "%%%d[^\n]", (unsigned int) length);
 
     int ret = scanf(format, str);
